Exception handling

+++ src/main/java/com/eduardodev/dscommerce/services/ProductService.java
@@ -1,16 +1,19 @@
 package com.eduardodev.dscommerce.services;
 
-import java.util.Optional;
-
+import com.eduardodev.dscommerce.services.exceptions.DataBaseException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.eduardodev.dscommerce.dto.ProductDTO;
 import com.eduardodev.dscommerce.entities.Product;
 import com.eduardodev.dscommerce.repositories.ProductRepository;
+import com.eduardodev.dscommerce.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class ProductService {
@@ -20,8 +23,8 @@ public class ProductService {
 	
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Optional<Product> result = repository.findById(id);
-		Product product = result.get();
+		Product product = repository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
 		ProductDTO dto = new ProductDTO(product);
 		return dto;
 	}
@@ -32,27 +35,40 @@ public class ProductService {
 	}
 	@Transactional
 	public ProductDTO insert(ProductDTO dto) {
-		
-		Product entity = new Product();
-		copyDtoToEntity(entity, dto);
-		entity = repository.save(entity);
-		return new ProductDTO(entity);
+			Product entity = new Product();
+			copyDtoToEntity(entity, dto);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);
 	}
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product entity = repository.getReferenceById(id);
-		copyDtoToEntity(entity, dto);
-		entity = repository.save(entity);
-		return new ProductDTO(entity);
+
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDtoToEntity(entity, dto);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);
+		}
+		catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não entrado");
+		}
+	}
+	@Transactional(propagation = Propagation.SUPPORTS)
+	public void delete(Long id) {
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não entrado");
+		}
+		try {
+			repository.deleteById(id);
+		}
+		catch (DataIntegrityViolationException e) {
+			throw new DataBaseException("Falha de integridade referencial");
+		}
 	}
 	private void copyDtoToEntity(Product entity, ProductDTO dto) {
 		entity.setName(dto.getName());
 		entity.setDescription(dto.getDescription());
 		entity.setPrice(dto.getPrice());
 		entity.setImgUrl(dto.getImgUrl());
-	}
-	@Transactional
-	public void delete(Long id) {
-		repository.deleteById(id);
 	}
 }
